+++ src/components/DateInput.vue
@@ -1,18 +1,36 @@
 <template>
-  <SeamlessInput
-    :value="dateValue"
+  <NcDatetimePicker
+    class="date-input"
+    v-model="dateValue"
+    :append-to-body="true"
     :placeholder="placeholder"
-    @value-changed="handleDateValueChanged"
+    :formatter="formatter"
+    :disabled="!editable"
   />
 </template>
+
+<style>
+
+.date-input.mx-datepicker .mx-input-wrapper .mx-input {
+  background: transparent !important;
+  border: 2px solid transparent !important;
+  box-shadow: none;
+}
+
+.date-input.mx-datepicker:not(.disabled) .mx-input-wrapper .mx-input:hover,
+.date-input.mx-datepicker .mx-input-wrapper .mx-input:focus {
+  border-color: var(--color-primary-element) !important;
+}
+
+</style>
 
 <script setup lang="ts">
 
   import dayjs from 'dayjs';
 
-  import { computed, onMounted, ref, watch } from 'vue';
+  import { computed } from 'vue';
 
-  import SeamlessInput from './SeamlessInput.vue';
+  import NcDatetimePicker from '@nextcloud/vue/dist/Components/NcDatetimePicker';
 
   const props = defineProps({
     date: {
@@ -22,32 +40,33 @@
     placeholder: {
       type: String,
       default: ''
+    },
+    editable: {
+      type: Boolean,
+      default: true
     }
   });
 
-  const emit = defineEmits(['date-changed'])
+  const emit = defineEmits(['date-changed']);
 
-  const dateValue = ref('');
-
-  const formattedDate = computed(() => {
-    return dayjs(props.date).format('L');
-  });
-
-  watch(formattedDate, () => {
-    dateValue.value = formattedDate.value;
-  })
-
-  function handleDateValueChanged(newDateValue: string) {
-    dateValue.value = formattedDate.value;
-
-    const newDate = dayjs(newDateValue, 'L').toDate();
-    if (!Number.isNaN(newDate.valueOf())) {
+  const dateValue = computed({
+    get() {
+      return props.date;
+    },
+    set(newDateValue) {
+      const newDate = newDateValue ?? dayjs().toDate();
       emit('date-changed', newDate);
     }
-  }
-
-  onMounted(() => {
-    dateValue.value = formattedDate.value;
   });
+
+  const formatter = {
+    stringify: (date: Date) => dayjs(date).format('L'),
+    parse: (value: string) => getDateFromValue(value)
+  };
+
+  function getDateFromValue(value: string): Date {
+    const date = ['L', 'D', 'DD'].map(f => dayjs(value, f)).find(djs => djs.isValid());
+    return (date ?? dayjs()).toDate();
+  }
 
 </script>